+++ packages/expo-modules-core/ios/Swift/ModuleHolder.swift
@@ -4,30 +4,71 @@ import Dispatch
  Holds a reference to the module instance and caches its definition.
  */
 public class ModuleHolder {
-  private(set) var module: AnyModule
+  /**
+   Lazy-loaded instance of the module. Created from module's definition.
+   */
+  private(set) var module: AnyModule?
 
-  private(set) lazy var definition: ModuleDefinition = module.definition()
+  /**
+   A weak reference to the app context.
+   */
+  private(set) weak var appContext: AppContext?
 
-  var name: String {
-    return definition.name ?? String(describing: type(of: module))
+  /**
+   Caches the definition of the module type.
+   */
+  let definition: ModuleDefinition
+
+  /**
+   Shorthand form of `definition.name`.
+   */
+  var name: String { definition.name }
+
+  init(appContext: AppContext, definition: ModuleDefinition) {
+    self.appContext = appContext
+    self.definition = definition
   }
 
-  init(module: AnyModule) {
-    self.module = module
+  /**
+   Creates (if needed) and returns the module based on associated definition and then returns it.
+   */
+  func getInstance() throws -> AnyModule? {
+    guard let appContext = appContext else {
+      throw AppContext.DeallocatedAppContextError()
+    }
+    if module == nil, let moduleType = definition.type {
+      module = moduleType.init(appContext: appContext)
+      post(event: .moduleCreate)
+    }
+    return module
+  }
 
-    post(event: .moduleCreate)
+  /**
+   Whether held module is or would be (instance may not be created yet) of given type.
+   */
+  func isOfType<ModuleType>(_ type: ModuleType.Type) -> Bool {
+    return definition.type is ModuleType.Type
   }
 
   // MARK: Calling methods
 
   func call(method methodName: String, args: [Any?], promise: Promise) {
-    if let method = definition.methods[methodName] {
-      let queue = method.queue ?? DispatchQueue.global(qos: .default)
-      queue.async {
-        method.call(args: args, promise: promise)
+    do {
+      guard let module = try getInstance() else {
+        throw ModuleNotFoundError(moduleName: self.name)
       }
-    } else {
-      promise.reject(MethodNotFoundError(methodName: methodName, moduleName: self.name))
+      guard let method = definition.methods[methodName] else {
+        throw MethodNotFoundError(methodName: methodName, moduleName: self.name)
+      }
+      let queue = method.queue ?? DispatchQueue.global(qos: .default)
+
+      queue.async {
+        method.call(module: module, args: args, promise: promise)
+      }
+    } catch let error as CodedError {
+      promise.reject(error)
+    } catch {
+      promise.reject(UnexpectedError(error))
     }
   }
 
@@ -50,13 +91,13 @@ public class ModuleHolder {
 
   func post(event: EventName) {
     listeners(forEvent: event).forEach {
-      $0.call(nil)
+      try? $0.call(module, nil)
     }
   }
 
   func post<PayloadType>(event: EventName, payload: PayloadType?) {
     listeners(forEvent: event).forEach {
-      $0.call(payload)
+      try? $0.call(module, payload)
     }
   }
 
@@ -67,6 +108,13 @@ public class ModuleHolder {
   }
 
   // MARK: Errors
+
+  struct ModuleNotFoundError: CodedError {
+    let moduleName: String
+    var description: String {
+      "Cannot find module `\(moduleName)`"
+    }
+  }
 
   struct MethodNotFoundError: CodedError {
     let methodName: String