+++ packages/expo-modules-core/ios/Swift/Modules/ModuleDefinitionComponents.swift
@@ -7,26 +7,47 @@ import UIKit
  will be implemented in the future.
  */
 extension AnyModule {
+  // MARK: - Module name
+
   /**
    Sets the name of the module that is exported to the JavaScript world.
    */
-  public func name(_ name: String) -> AnyDefinition {
+  public static func name(_ name: String) -> AnyDefinition {
     return ModuleNameDefinition(name: name)
   }
+
+  // MARK: - Constants
 
   /**
    Definition function setting the module's constants to export.
    */
-  public func constants(_ closure: () -> [String : Any?]) -> AnyDefinition {
+  public static func constants(_ closure: () -> [String : Any?]) -> AnyDefinition {
     return ConstantsDefinition(constants: closure())
+  }
+
+  // MARK: - Methods
+
+  /**
+   Factory function for methods without the module instance and arguments.
+   */
+  public static func method<R>(
+    _ name: String,
+    _ closure: @escaping () -> R
+  ) -> AnyMethod {
+    return ConcreteMethod(
+      name,
+      argTypes: [],
+      closure,
+      detached: true
+    )
   }
 
   /**
    Factory function for methods without arguments.
    */
-  public func method<R>(
+  public static func method<R>(
     _ name: String,
-    _ closure: @escaping () -> R
+    _ closure: @escaping (Self) -> R
   ) -> AnyMethod {
     return ConcreteMethod(
       name,
@@ -38,9 +59,9 @@ extension AnyModule {
   /**
    Factory function for methods with one argument.
    */
-  public func method<R, A0: AnyMethodArgument>(
+  public static func method<R, A0: AnyMethodArgument>(
     _ name: String,
-    _ closure: @escaping (A0) -> R
+    _ closure: @escaping (Self, A0) -> R
   ) -> AnyMethod {
     return ConcreteMethod(
       name,
@@ -52,9 +73,9 @@ extension AnyModule {
   /**
    Factory function for methods with 2 arguments.
    */
-  public func method<R, A0: AnyMethodArgument, A1: AnyMethodArgument>(
+  public static func method<R, A0: AnyMethodArgument, A1: AnyMethodArgument>(
     _ name: String,
-    _ closure: @escaping (A0, A1) -> R
+    _ closure: @escaping (Self, A0, A1) -> R
   ) -> AnyMethod {
     return ConcreteMethod(
       name,
@@ -66,9 +87,9 @@ extension AnyModule {
   /**
    Factory function for methods with 3 arguments.
    */
-  public func method<R, A0: AnyMethodArgument, A1: AnyMethodArgument, A2: AnyMethodArgument>(
+  public static func method<R, A0: AnyMethodArgument, A1: AnyMethodArgument, A2: AnyMethodArgument>(
     _ name: String,
-    _ closure: @escaping (A0, A1, A2) -> R
+    _ closure: @escaping (Self, A0, A1, A2) -> R
   ) -> AnyMethod {
     return ConcreteMethod(
       name,
@@ -80,9 +101,9 @@ extension AnyModule {
   /**
    Factory function for methods with 4 arguments.
    */
-  public func method<R, A0: AnyMethodArgument, A1: AnyMethodArgument, A2: AnyMethodArgument, A3: AnyMethodArgument>(
+  public static func method<R, A0: AnyMethodArgument, A1: AnyMethodArgument, A2: AnyMethodArgument, A3: AnyMethodArgument>(
     _ name: String,
-    _ closure: @escaping (A0, A1, A2, A3) -> R
+    _ closure: @escaping (Self, A0, A1, A2, A3) -> R
   ) -> AnyMethod {
     return ConcreteMethod(
       name,
@@ -94,9 +115,9 @@ extension AnyModule {
   /**
    Factory function for methods with 5 arguments.
    */
-  public func method<R, A0: AnyMethodArgument, A1: AnyMethodArgument, A2: AnyMethodArgument, A3: AnyMethodArgument, A4: AnyMethodArgument>(
+  public static func method<R, A0: AnyMethodArgument, A1: AnyMethodArgument, A2: AnyMethodArgument, A3: AnyMethodArgument, A4: AnyMethodArgument>(
     _ name: String,
-    _ closure: @escaping (A0, A1, A2, A3, A4) -> R
+    _ closure: @escaping (Self, A0, A1, A2, A3, A4) -> R
   ) -> AnyMethod {
     return ConcreteMethod(
       name,
@@ -108,9 +129,9 @@ extension AnyModule {
   /**
    Factory function for methods with 6 arguments.
    */
-  public func method<R, A0: AnyMethodArgument, A1: AnyMethodArgument, A2: AnyMethodArgument, A3: AnyMethodArgument, A4: AnyMethodArgument, A5: AnyMethodArgument>(
+  public static func method<R, A0: AnyMethodArgument, A1: AnyMethodArgument, A2: AnyMethodArgument, A3: AnyMethodArgument, A4: AnyMethodArgument, A5: AnyMethodArgument>(
     _ name: String,
-    _ closure: @escaping (A0, A1, A2, A3, A4, A5) -> R
+    _ closure: @escaping (Self, A0, A1, A2, A3, A4, A5) -> R
   ) -> AnyMethod {
     return ConcreteMethod(
       name,
@@ -122,9 +143,9 @@ extension AnyModule {
   /**
    Factory function for methods with 7 arguments.
    */
-  public func method<R, A0: AnyMethodArgument, A1: AnyMethodArgument, A2: AnyMethodArgument, A3: AnyMethodArgument, A4: AnyMethodArgument, A5: AnyMethodArgument, A6: AnyMethodArgument>(
+  public static func method<R, A0: AnyMethodArgument, A1: AnyMethodArgument, A2: AnyMethodArgument, A3: AnyMethodArgument, A4: AnyMethodArgument, A5: AnyMethodArgument, A6: AnyMethodArgument>(
     _ name: String,
-    _ closure: @escaping (A0, A1, A2, A3, A4, A5, A6) -> R
+    _ closure: @escaping (Self, A0, A1, A2, A3, A4, A5, A6) -> R
   ) -> AnyMethod {
     return ConcreteMethod(
       name,
@@ -136,9 +157,9 @@ extension AnyModule {
   /**
    Factory function for methods with 8 arguments.
    */
-  public func method<R, A0: AnyMethodArgument, A1: AnyMethodArgument, A2: AnyMethodArgument, A3: AnyMethodArgument, A4: AnyMethodArgument, A5: AnyMethodArgument, A6: AnyMethodArgument, A7: AnyMethodArgument>(
+  public static func method<R, A0: AnyMethodArgument, A1: AnyMethodArgument, A2: AnyMethodArgument, A3: AnyMethodArgument, A4: AnyMethodArgument, A5: AnyMethodArgument, A6: AnyMethodArgument, A7: AnyMethodArgument>(
     _ name: String,
-    _ closure: @escaping (A0, A1, A2, A3, A4, A5, A6, A7) -> R
+    _ closure: @escaping (Self, A0, A1, A2, A3, A4, A5, A6, A7) -> R
   ) -> AnyMethod {
     return ConcreteMethod(
       name,
@@ -147,66 +168,122 @@ extension AnyModule {
     )
   }
 
+  // MARK: - `onCreate`
+
   /**
    Creates module's lifecycle listener that is called right after module initialization.
    */
-  public func onCreate(_ closure: @escaping () -> Void) -> AnyDefinition {
+  public static func onCreate(_ closure: @escaping () -> Void) -> AnyDefinition {
     return EventListener(.moduleCreate, closure)
+  }
+
+  /**
+   Creates module's lifecycle listener that is called right after module initialization.
+   */
+  public static func onCreate(_ closure: @escaping (Self) -> Void) -> AnyDefinition {
+    return EventListener(.moduleCreate, closure)
+  }
+
+  // MARK: - `onDestroy`
+
+  /**
+   Creates module's lifecycle listener that is called when the module is about to be deallocated.
+   */
+  public static func onDestroy(_ closure: @escaping () -> Void) -> AnyDefinition {
+    return EventListener(.moduleDestroy, closure)
   }
 
   /**
    Creates module's lifecycle listener that is called when the module is about to be deallocated.
    */
-  public func onDestroy(_ closure: @escaping () -> Void) -> AnyDefinition {
+  public static func onDestroy(_ closure: @escaping (Self) -> Void) -> AnyDefinition {
     return EventListener(.moduleDestroy, closure)
+  }
+
+  // MARK: - `onAppContextDestroys`
+
+  /**
+   Creates module's lifecycle listener that is called when the app context owning the module is about to be deallocated.
+   */
+  public static func onAppContextDestroys(_ closure: @escaping () -> Void) -> AnyDefinition {
+    return EventListener(.appContextDestroys, closure)
   }
 
   /**
    Creates module's lifecycle listener that is called when the app context owning the module is about to be deallocated.
    */
-  public func onAppContextDestroys(_ closure: @escaping () -> Void) -> AnyDefinition {
+  public static func onAppContextDestroys(_ closure: @escaping (Self) -> Void) -> AnyDefinition {
     return EventListener(.appContextDestroys, closure)
+  }
+
+  // MARK: - `onAppEntersForeground`
+
+  /**
+   Creates a listener that is called when the app is about to enter the foreground mode.
+   */
+  public static func onAppEntersForeground(_ closure: @escaping () -> Void) -> AnyDefinition {
+    return EventListener(.appEntersForeground, closure)
   }
 
   /**
    Creates a listener that is called when the app is about to enter the foreground mode.
    */
-  public func onAppEntersForeground(_ closure: @escaping () -> Void) -> AnyDefinition {
+  public static func onAppEntersForeground(_ closure: @escaping (Self) -> Void) -> AnyDefinition {
     return EventListener(.appEntersForeground, closure)
+  }
+
+  // MARK: - `onAppBecomesActive`
+
+  /**
+   Creates a listener that is called when the app becomes active again.
+   */
+  public static func onAppBecomesActive(_ closure: @escaping () -> Void) -> AnyDefinition {
+    return EventListener(.appBecomesActive, closure)
   }
 
   /**
    Creates a listener that is called when the app becomes active again.
    */
-  public func onAppBecomesActive(_ closure: @escaping () -> Void) -> AnyDefinition {
+  public static func onAppBecomesActive(_ closure: @escaping (Self) -> Void) -> AnyDefinition {
     return EventListener(.appBecomesActive, closure)
+  }
+
+  // MARK: - `onAppEntersBackground`
+
+  /**
+   Creates a listener that is called when the app enters the background mode.
+   */
+  public static func onAppEntersBackground(_ closure: @escaping () -> Void) -> AnyDefinition {
+    return EventListener(.appEntersBackground, closure)
   }
 
   /**
    Creates a listener that is called when the app enters the background mode.
    */
-  public func onAppEntersBackground(_ closure: @escaping () -> Void) -> AnyDefinition {
+  public static func onAppEntersBackground(_ closure: @escaping (Self) -> Void) -> AnyDefinition {
     return EventListener(.appEntersBackground, closure)
   }
-}
 
-/**
- Creates the view manager definition that scopes other view-related definitions.
- */
-public func viewManager(@ViewManagerDefinitionBuilder _ closure: @escaping () -> ViewManagerDefinition) -> AnyDefinition {
-  return closure()
-}
+  // MARK: - View Manager
 
-/**
- Defines the factory creating a native view when the module is used as a view.
- */
-public func view(_ closure: @escaping () -> UIView) -> AnyDefinition {
-  return ViewFactory(closure)
-}
+  /**
+   Creates the view manager definition that scopes other view-related definitions.
+   */
+  public static func viewManager(@ViewManagerDefinitionBuilder _ closure: @escaping () -> ViewManagerDefinition) -> AnyDefinition {
+    return closure()
+  }
 
-/**
- Creates a view prop that defines its name and setter.
- */
-public func prop<ViewType: UIView, PropType>(_ name: String, _ setter: @escaping (ViewType, PropType) -> Void) -> AnyDefinition {
-  return ConcreteViewProp(name, setter)
+  /**
+   Defines the factory creating a native view when the module is used as a view.
+   */
+  public static func view(_ closure: @escaping () -> UIView) -> AnyDefinition {
+    return ViewFactory(closure)
+  }
+
+  /**
+   Creates a view prop that defines its name and setter.
+   */
+  public static func prop<ViewType: UIView, PropType>(_ name: String, _ setter: @escaping (ViewType, PropType) -> Void) -> AnyDefinition {
+    return ConcreteViewProp(name, setter)
+  }
 }